poc/00-gpt-api/requirements.txt
pandas
numpy
openai>=1.2.0
ipykernel
jupyterlab
python-dotenv